Break examples/test.rs into several tighter, more focused examples.
Additionally, added more comments and explanations.
